refactor: move images apis
