+++ lambdas/src/apis/images/controllers/images.ts
