+++ lambdas/src/controllers/routes.ts
@@ -11,6 +11,9 @@ import { getWearablesByOwnerHandler, getWearablesHandler } from '../apis/collect
 import { OffChainWearablesManager } from '../apis/collections/off-chain/OffChainWearablesManager'
 import { getContents, getInfo, getScenes } from '../apis/content-v2/controllers/translator'
 import { getCatalystServersList, getDenylistedNamesList, getPOIsList } from '../apis/contracts/controllers/contracts'
+import { validateSignature } from '../apis/crypto/controllers/crypto'
+import { hotScenes, realmsStatus } from '../apis/explore/controllers/explore'
+import { getResizedImage } from '../apis/images/controllers/images'
 import { Bean, Environment, EnvironmentConfig } from '../Environment'
 import { createFetchComponent } from '../ports/fetcher'
 import { createThirdPartyAssetFetcher, ThirdPartyAssetFetcher } from '../ports/third-party/third-party-fetcher'
@@ -18,9 +21,6 @@ import { DAOCache } from '../service/dao/DAOCache'
 import { SmartContentClient } from '../utils/SmartContentClient'
 import { SmartContentServerFetcher } from '../utils/SmartContentServerFetcher'
 import { TheGraphClient } from '../utils/TheGraphClient'
-import { validateSignature } from './handlers/crypto/handlers'
-import { hotScenes, realmsStatus } from './handlers/explorer/handlers'
-import { getResizedImage } from './handlers/images/handlers'
 import { EmotesOwnership } from './handlers/profiles/EmotesOwnership'
 import { EnsOwnership } from './handlers/profiles/EnsOwnership'
 import { createProfileHandler, getIndividualProfileById, getProfilesById } from './handlers/profiles/handlers'
